% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mining_esg_index.R
\name{plot_mining_esg_index}
\alias{plot_mining_esg_index}
\title{Plot Mining ESG Composite Index}
\usage{
plot_mining_esg_index(index_data)
}
\arguments{
\item{index_data}{data.table returned by generate_mining_esg_index()}
}
\value{
ggplot object
}
\description{
Creates a ranked horizontal bar chart of composite ESG risk.
}
